rename colber_embeddings file
